microservice/requirements.txt
cffi==1.15.1
defusedxml==0.7.1
pandas==1.5.3
pillow==9.4.0
scipy==1.10.0
tensorflow-intel
tensorflow
Flask==2.2.2